Convert int to double

+++ HM5/src/com/company/TestClass2.java
@@ -4,21 +4,21 @@ import java.util.Arrays;
 
 public class TestClass2
 {
-    MySortedList<Integer> list = new MySortedList<>();
+    MySortedList<Double> list = new MySortedList<>();
 
     public void run()
     {
         //Add elements
-        list.add(1);
-        list.add(100);
-        list.add(2);
-        list.add(4);
-        list.add(0);
-        list.add(8);
-        list.add(33);
-        list.add(44);
-        list.add(5);
-        list.add(7);
+        list.add(Double.valueOf(1));
+        list.add(Double.valueOf(100));
+        list.add(Double.valueOf(2));
+        list.add(Double.valueOf(4));
+        list.add(Double.valueOf(0));
+        list.add(Double.valueOf(8));
+        list.add(Double.valueOf(33));
+        list.add(Double.valueOf(44));
+        list.add(Double.valueOf(5));
+        list.add(Double.valueOf(7));
         System.out.println(list);
 
         //Remove elements from index
